update status, add stop method and abstract estado
